Move contents of read and write modules into dakota.py
